Parse only the digits from the tesseract version string.
The current tesseract 4.0 version is still in alpha and returns a
respective version string [0]. This breaks the existing get_version
function as it expects integer values only. To work around it we simply
only take the starting digits of the version and return these.

Note that this commit only allows the version check to pass. This in
itself does not say anything about the support of tesseract 4.0.

[0]
$ tesseract -v
$ tesseract 4.00.00alpha